+++ Function.h
@@ -4,7 +4,7 @@ class Function
 public:
 	~Function() = default;
 	virtual double Execute() = 0;
-	virtual double Execute(int x) = 0;	
-	virtual double Execute(int x, int y) = 0;
+	virtual double Execute(double x) = 0;	
+	virtual double Execute(double x, double y) = 0;
 };
