+++ Min.cpp
@@ -3,7 +3,7 @@
 Min::Min() {}
 Min::~Min() {}
 
-double Min::Execute(int x, int y)
+double Min::Execute(double x, double y)
 {
 	if (x <= y)
 		return x;
